fix(rev-254): audit latent, subsystem group — int32 wraps, bzip2 signedness, ctor order, audio caps
- wordfilter-B-01: FirstFragmentID accumulator wraps to int32 per
  assignment (Java WordFilter.java:1053-1071 32-bit int; only length-3
  fragments are reachable today).
- io-bzip2-02: Huffman code-length storage retyped [][]byte -> [][]int8
  (Java signed byte[][]): min/max scan now compares signed,
  CreateDecodeTables perm fill sign-extends, and the base count indexes
  with Java's promote-then-add (negative index panics faithfully).
  Divergent only on malformed streams (well-formed lengths <= 20).
- config-B-04: IfType.CacheModel key now int64(int32((typ<<16)+id)) —
  Java widens AFTER the 32-bit add (IfType.java:500), consistent with
  LoadModel's already-correct key.
- dash3d-collision-01: NewCollisionMap maps 2nd param -> SizeX, 1st ->
  SizeZ per CollisionMap.java:23-28; masked until now by always-square
  104x104 callers.
- signlink-02: SaveWave rejects (returns false) and SaveMidi silently
  drops payloads over 2,000,000 bytes, restoring the signlink.java
  publisher caps (255-257/281-284 @2e62978); the savereq busy-slot half
  stays unported (documented in-memory-seam decision).

Gates: build, vet, test -race (31 pkgs), gofmt, wasm build, golangci-lint 0.

+++ pkg/jagex2/config/iftype/iftype.go
@@ -413,7 +413,9 @@ func (c *IfType) LoadModel(arg0 int, arg1 int, localPlayer *playerentity.ClientP
 func CacheModel(m *model.Model, id int, typ int) {
 	ModelCache.Clear()
 	if m != nil && typ != 4 {
-		ModelCache.Put((int64(typ)<<16)+int64(id), m)
+		// Java: (long)((arg3 << 16) + arg1) — the add runs in 32-bit int and
+		// wraps before widening, like LoadModel's key (audit config-B-04).
+		ModelCache.Put(int64(int32((typ<<16)+id)), m)
 	}
 }
